php finished

+++ admin/phpPro/delete.php
@@ -1,12 +1,18 @@
 <?php
 
     include "../../functions/functions.php";
+    session_start();
 
-    if(isset($_GET["id"])){
-        DeleteBlogs(intval($_GET["id"]));
-        MessageFunc("Silindi","danger","../../blogs.php");
+    if(isset($_SESSION["isLogin"]) && $_SESSION["isAdmin"] == "true"){
+        if(isset($_GET["id"])){
+            DeleteBlogs(intval($_GET["id"]));
+            MessageFunc("Silindi","danger","../../blogs.php");
+        }else{
+            header("Location:../index.php");
+            exit();
+        }
     }else{
-        header("Location:../index.php");
+        header("Location:../../index.php");
         exit();
     }
