+++ admin/phpPro/deleteUser.php
@@ -2,12 +2,21 @@
 
     include "../../functions/functions.php";
 
-    if(isset($_GET["id"])){
-        DeleteUsers($_GET["id"]);
-        MessageFunc("Kullanıcı silindi","danger","../../users.php");
+    session_start();
+
+    if(isset($_SESSION["isLogin"]) && $_SESSION["isAdmin"] == "true"){
+
+        
+        if(isset($_GET["id"])){
+            DeleteUsers($_GET["id"]);
+            MessageFunc("Kullanıcı silindi","danger","../../users.php");
+        }else{
+            header("Location:../users.php");
+            exit();
+        }
     }else{
-        header("Location:../users.php");
+        header("Location:../../index.php");
         exit();
     }
-
+        
 ?>
